Update README.md
Hi, About my CA, I tried a few methods and different versions of different nodes, and also downloaded two programs that could not make docker and postgres, pgadmin4 connected. 

After the cloud server (terminal) is established remotely. Use pycharm to upload items to a remote server and add configurations. After the configuration is complete, there is an error uploading the local project to the remote server, which cannot be uploaded successfully and was not resolved successfully.

About serviceworker, I used cache api. First define a callback for the install event and determine which files need to be cached. Turn on a cache in call back Cache our files Determines whether all resources are cached. After you install the service worker, consider returning your cached request. When the user browses the website, the browser tries to re-download the script file for the service worker in the background. By comparison, as long as the files on the server are one byte different from the local files, the file is considered new. The updated service worker then starts and triggers the install event. When we returned the cache request, we defined the feed event in event.respondWith, and we passed in a promise, generated by caches.match. caches.match looks for the request to see if the request hits the cache previously set by the service worker. 

Deploy the rest framework inside Login. First create the user table model. Then create a new serializer: Create a new serializers.py. Gets a list of data in the view and creates a class view of the data. A auth.py is then established for user authentication. Restricting access to user conditions allows logins to obtain a list of users through GET. 

The first step is to configure the settings.py of the geoserver leaflet project so that html can be accessed. First, in the settings.py, add the templates path. Add a static file path, static_files, and here are the css and js. Add in the body label, set its style in the style label, and add a statement to the script label. Load the layer resources inside the data as base layers. The shp data can be converted to get geojson data. The Feature Turn json tool in the arcgis toolbox exports the EsriJson format, and arcgis pro supports the export geojson format.